Changed mock difinition Added the persist engine concept New variable parsing approach Added new variables
