+++ config/agenda.yaml
@@ -1,0 +1,19 @@
+request: 
+  method: GET
+  path: /agenda/show.json
+  queryStringParameters: 		
+    user_id:
+      - 123
+response:
+  statusCode: 200
+  headers:
+    Content-Type: 
+      - "application/json"
+  body: >
+    {  
+      "id":"{{request.query.user_id}}",
+      "first_name":"{{fake.FirstName}}",
+      "last_name":"{{fake.LastName}}",
+      "gender":"{{fake.Gender}}",
+      "email":"{{fake.EmailAddress}}"
+    }
