--- config/amqp-post.json
@@ -1,26 +1,0 @@
-{
-    "request": {
-        "method": "POST",
-        "path": "/amqp/*",
-        "body": "*"
-    },
-	"response": {
-        "statusCode": 202,
-        "headers": {
-			"Content-Type":["application/json"]
-		},
-		"body": "{{ request.body }}",
-        "bodyAppend": "{ \"id\": {{request.url./amqp/(?P<value>\\d+)}}, \"city\": \"{{ fake.City }}\" }"
-    },
-    "persist" : {
-        "name" : "/user-{{request.url./amqp/(?P<value>\\d+)}}.json",
-		"amqp": {
-            "url": "amqp://guest:guest@localhost:5672/myVHost",
-            "body": "{{ response.body }}",
-            "delay": 2,
-            "exchange": "myExchange",
-            "type": "MockType",
-            "correlationId": "9782b88f-0c6e-4879-8c23-4699785e6a95"
-        }
-	}
-}
