--- config/hello_again.json
@@ -1,24 +1,0 @@
-{
-	"request": {
-		"method": "GET",
-		"path": "/hello*",
-		"queryStringParameters": {
-			"name": ["world"]
-		},
-		"headers": {},
-		"cookies": {
-			"visit" : "true"
-		},
-		"body": ""
-	},
-	"response": {
-		"statusCode": 200,
-		"headers": {
-			"Content-Type":["application/json"]
-		},
-		"body": "Hello AGAIN! my name is {{fake.FirstName}}!"
-	},
-	"control":{
-		"priority":1
-	}
-}
