--- config/user_set_name.json
@@ -1,20 +1,0 @@
-{
-	"request": {
-		"method": "PATCH",
-		"path": "users/*",
-		"headers": {
-			"Content-Type":["application/json"]
-		},
-		"body": "{\"first_name\": \"Jimmy Jump\"}"
-	},
-	"response": {
-		"statusCode": 200,
-		"headers": {
-			"Content-Type":["application/json"]
-		},
-		"body": "{\"result\": \"ok\"}"
-	},
-	"control":{
-		"delay":1
-	}
-}
