+++ persist/entity_persister.go
@@ -1,0 +1,8 @@
+package persist
+
+type EntityPersister interface {
+	Read(name string) (string, error)
+	Write(name, content string) error
+	Delete(name string) error
+	GetName() string
+}
