+++ vars/filler.go
@@ -1,0 +1,7 @@
+package vars
+
+import "github.com/jmartin82/mmock/definition"
+
+type Filler interface {
+	Fill(m *definition.Mock, input string) string
+}
